Some machines are having trouble compiling our SDK, particularly libAVSUtils.so. The issue seems to be that the linker is encountering an error finding a reference to “pthread_create”.
AVSUtils/src/Logger/ThreadIdAsString.cpp, has a static thread_local variable and includes <thread>, and since this file is used indirectly from the ExampleLogger, this issue is showing up as an issue in ExampleLogger.

Currently, ACL and AuthDelegate both find the Threads package and link against in their src/CMakeLists.txt file, so we don't see such issues when creating those libraries. However, AVSUtils does not do this. The fix for this particular issue (https://github.com/alexa/alexa-client-sdk/issues/3) is to find the Threads package and link against it in the AVSUtils library as well, which this commit does. This commit also pulls in the Threads package and links against it in other areas that include <thread> so that we don’t run into this issue down the road by mistake.