GstreamerMediaPlayer: Fix callback function use-after-free
Allocate a copy of the callback function and promise, and keep it until it
was executed. This fixes a problem where the callback function could
be freed as soon as the callback function sets the promise value and
the thread waiting on the promise exits the function immediately, freeing
the closure data while it is still executing.